Update from original

+++ xbmc/filesystem/IFileTypes.h
@@ -1,7 +1,6 @@
-
 /*
- *      Copyright (C) 2005-2012 Team XBMC
- *      http://www.xbmc.org
+ *      Copyright (C) 2005-2013 Team XBMC
+ *      http://xbmc.org
  *
  *  This Program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -18,15 +17,18 @@
  *  <http://www.gnu.org/licenses/>.
  *
  */
+
 #pragma once
+
+#include <stdint.h>
 
 namespace XFILE
 {
 
 struct SNativeIoControl
 {
-  int   request;
-  void* param;
+  unsigned long int   request;
+  void*               param;
 };
 
 struct SCacheStatus
